fixed suggested issues
